control.in is correctly reading in the inputs
All data structures are coded for inputs, control.in properly reads in
the input but there is a bug that needs to be fixed so it will print all
values out properly. This will be fixed later today and hopefully the
rest of the classes will be implemented also. You can use the .sh file
to compile all files.